Fix build issues on riscv64 by using npm from Ubuntu
The node snap is not available for riscv64 and that caused build
failures on that architecture. So the snapcraft.yaml file now installs
npm from the Ubuntu archive instead using snapcraft advanced grammar.

Due to this, the snap is built using different node/npm versions on
riscv64 (18.19.1 at the time of this commit) than amd64 and arm64 which
use the node snap from the 22/stable channel.